rrpage: use Linux vDSO soname
AddressSanitizer complains if it isn't the first shared object in the
link map besides the vDSO. It checks for the vDSO by testing whether
the DT_SONAME of the object starts with 'linux-'. With the
introduction of librrpage, this check no longer passes and causes
ASan-instrumented programs to abort when being recorded by rr.

This patch alters the linker arguments used for producing the
librrpage dummy vDSO such that its SONAME is 'linux-vdso.so.1'.